cleaned up the code

+++ app/llama_model.py
@@ -3,6 +3,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+
 class RMSNorm(nn.Module):
     def __init__(self, dim, eps=1e-6):
         super().__init__()
@@ -13,34 +14,43 @@ class RMSNorm(nn.Module):
         norm = x.norm(2, dim=-1, keepdim=True)
         return x * torch.rsqrt(norm.pow(2) + self.eps) * self.weight
 
+
 class RotaryEmbedding(nn.Module):
     def __init__(self, dim, max_position_embeddings=2048, base=10000, device=None):
         super().__init__()
         inv_freq = 1.0 / (base ** (torch.arange(0, dim, 2).float().to(device) / dim))
         self.register_buffer("inv_freq", inv_freq)
         self.max_seq_len_cached = max_position_embeddings
-        t = torch.arange(self.max_seq_len_cached, device=self.inv_freq.device).type_as(self.inv_freq)
+        t = torch.arange(self.max_seq_len_cached, device=self.inv_freq.device).type_as(
+            self.inv_freq)
         freqs = torch.einsum("i,j->ij", t, self.inv_freq)
         emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer("cos_cached", emb.cos()[None, None, :, :], persistent=False)
-        self.register_buffer("sin_cached", emb.sin()[None, None, :, :], persistent=False)
+        self.register_buffer("cos_cached", emb.cos()[None, None, :, :],
+                             persistent=False)
+        self.register_buffer("sin_cached", emb.sin()[None, None, :, :],
+                             persistent=False)
 
     def forward(self, x, seq_len=None):
         if seq_len > self.max_seq_len_cached:
             self.max_seq_len_cached = seq_len
-            t = torch.arange(self.max_seq_len_cached, device=self.inv_freq.device).type_as(self.inv_freq)
+            t = torch.arange(self.max_seq_len_cached,
+                             device=self.inv_freq.device).type_as(self.inv_freq)
             freqs = torch.einsum("i,j->ij", t, self.inv_freq)
             emb = torch.cat((freqs, freqs), dim=-1).to(x.device)
-            self.register_buffer("cos_cached", emb.cos()[None, None, :, :], persistent=False)
-            self.register_buffer("sin_cached", emb.sin()[None, None, :, :], persistent=False)
+            self.register_buffer("cos_cached", emb.cos()[None, None, :, :],
+                                 persistent=False)
+            self.register_buffer("sin_cached", emb.sin()[None, None, :, :],
+                                 persistent=False)
         return (
             self.cos_cached[:, :, :seq_len, ...].to(dtype=x.dtype),
             self.sin_cached[:, :, :seq_len, ...].to(dtype=x.dtype),
         )
 
+
 def rotate_half(x):
-    x1, x2 = x[..., : x.shape[-1] // 2], x[..., x.shape[-1] // 2 :]
+    x1, x2 = x[..., : x.shape[-1] // 2], x[..., x.shape[-1] // 2:]
     return torch.cat((-x2, x1), dim=-1)
+
 
 def apply_rotary_pos_emb(q, k, cos, sin, position_ids):
     cos = cos.squeeze(1).squeeze(0)  # [seq_len, dim]
@@ -51,15 +61,22 @@ def apply_rotary_pos_emb(q, k, cos, sin, position_ids):
     k_embed = (k * cos) + (rotate_half(k) * sin)
     return q_embed, k_embed
 
+
 class LlamaMLP(nn.Module):
     def __init__(self, config):
         super().__init__()
-        self.gate_proj = nn.Linear(config["dim"], int(config["ffn_dim_multiplier"] * config["dim"]), bias=False)
-        self.up_proj = nn.Linear(config["dim"], int(config["ffn_dim_multiplier"] * config["dim"]), bias=False)
-        self.down_proj = nn.Linear(int(config["ffn_dim_multiplier"] * config["dim"]), config["dim"], bias=False)
+        self.gate_proj = nn.Linear(config["dim"],
+                                   int(config["ffn_dim_multiplier"] * config["dim"]),
+                                   bias=False)
+        self.up_proj = nn.Linear(config["dim"],
+                                 int(config["ffn_dim_multiplier"] * config["dim"]),
+                                 bias=False)
+        self.down_proj = nn.Linear(int(config["ffn_dim_multiplier"] * config["dim"]),
+                                   config["dim"], bias=False)
 
     def forward(self, x):
         return self.down_proj(F.silu(self.gate_proj(x)) * self.up_proj(x))
+
 
 class LlamaAttention(nn.Module):
     def __init__(self, config):
@@ -77,14 +94,20 @@ class LlamaAttention(nn.Module):
     def forward(self, hidden_states, attention_mask=None, position_ids=None):
         bsz, q_len, _ = hidden_states.size()
 
-        query_states = self.q_proj(hidden_states).view(bsz, q_len, self.n_heads, self.head_dim).transpose(1, 2)
-        key_states = self.k_proj(hidden_states).view(bsz, q_len, self.n_kv_heads, self.head_dim).transpose(1, 2)
-        value_states = self.v_proj(hidden_states).view(bsz, q_len, self.n_kv_heads, self.head_dim).transpose(1, 2)
+        query_states = self.q_proj(hidden_states).view(bsz, q_len, self.n_heads,
+                                                       self.head_dim).transpose(1, 2)
+        key_states = self.k_proj(hidden_states).view(bsz, q_len, self.n_kv_heads,
+                                                     self.head_dim).transpose(1, 2)
+        value_states = self.v_proj(hidden_states).view(bsz, q_len, self.n_kv_heads,
+                                                       self.head_dim).transpose(1, 2)
 
         cos, sin = self.rotary_emb(value_states, seq_len=q_len)
-        query_states, key_states = apply_rotary_pos_emb(query_states, key_states, cos, sin, position_ids)
+        query_states, key_states = apply_rotary_pos_emb(query_states, key_states, cos,
+                                                        sin, position_ids)
 
-        attn_weights = torch.matmul(query_states, key_states.transpose(2, 3)) / math.sqrt(self.head_dim)
+        attn_weights = torch.matmul(query_states,
+                                    key_states.transpose(2, 3)) / math.sqrt(
+            self.head_dim)
 
         if attention_mask is not None:
             attn_weights = attn_weights + attention_mask
@@ -95,6 +118,7 @@ class LlamaAttention(nn.Module):
         attn_output = attn_output.transpose(1, 2).contiguous().view(bsz, q_len, -1)
 
         return self.o_proj(attn_output)
+
 
 class LlamaDecoderLayer(nn.Module):
     def __init__(self, config):
@@ -107,7 +131,8 @@ class LlamaDecoderLayer(nn.Module):
     def forward(self, hidden_states, attention_mask=None, position_ids=None):
         residual = hidden_states
         hidden_states = self.input_layernorm(hidden_states)
-        hidden_states = self.self_attn(hidden_states, attention_mask=attention_mask, position_ids=position_ids)
+        hidden_states = self.self_attn(hidden_states, attention_mask=attention_mask,
+                                       position_ids=position_ids)
         hidden_states = residual + hidden_states
 
         residual = hidden_states
@@ -117,6 +142,7 @@ class LlamaDecoderLayer(nn.Module):
 
         return hidden_states
 
+
 class CustomLlamaModel(nn.Module):
     def __init__(self, config):
         super().__init__()
@@ -125,14 +151,16 @@ class CustomLlamaModel(nn.Module):
         self.embed_dim = config["dim"]
 
         self.embed_tokens = nn.Embedding(config["vocab_size"], config["dim"])
-        self.layers = nn.ModuleList([LlamaDecoderLayer(config) for _ in range(config["n_layers"])])
+        self.layers = nn.ModuleList(
+            [LlamaDecoderLayer(config) for _ in range(config["n_layers"])])
         self.norm = RMSNorm(config["dim"], eps=config["norm_eps"])
 
     def forward(self, input_ids, attention_mask=None, position_ids=None):
         hidden_states = self.embed_tokens(input_ids)
 
         for layer in self.layers:
-            hidden_states = layer(hidden_states, attention_mask=attention_mask, position_ids=position_ids)
+            hidden_states = layer(hidden_states, attention_mask=attention_mask,
+                                  position_ids=position_ids)
 
         hidden_states = self.norm(hidden_states)
 
@@ -140,27 +168,31 @@ class CustomLlamaModel(nn.Module):
 
     def generate(self, input_ids, max_length, temperature=1.0, top_p=0.9):
         for _ in range(max_length - input_ids.shape[1]):
-            attention_mask = torch.ones(input_ids.shape, dtype=torch.long, device=input_ids.device)
+            attention_mask = torch.ones(input_ids.shape, dtype=torch.long,
+                                        device=input_ids.device)
             position_ids = attention_mask.long().cumsum(-1) - 1
-            
-            outputs = self(input_ids, attention_mask=attention_mask, position_ids=position_ids)
+
+            outputs = self(input_ids, attention_mask=attention_mask,
+                           position_ids=position_ids)
             next_token_logits = outputs[:, -1, :]
-            
+
             # Apply temperature
             next_token_logits = next_token_logits / temperature
-            
+
             # Apply top-p (nucleus) sampling
-            sorted_logits, sorted_indices = torch.sort(next_token_logits, descending=True)
+            sorted_logits, sorted_indices = torch.sort(next_token_logits,
+                                                       descending=True)
             cumulative_probs = torch.cumsum(F.softmax(sorted_logits, dim=-1), dim=-1)
             sorted_indices_to_remove = cumulative_probs > top_p
-            sorted_indices_to_remove[..., 1:] = sorted_indices_to_remove[..., :-1].clone()
+            sorted_indices_to_remove[..., 1:] = sorted_indices_to_remove[...,
+                                                :-1].clone()
             sorted_indices_to_remove[..., 0] = 0
             indices_to_remove = sorted_indices[sorted_indices_to_remove]
             next_token_logits[:, indices_to_remove] = float('-inf')
-            
+
             probs = F.softmax(next_token_logits, dim=-1)
             next_token = torch.multinomial(probs, num_samples=1).squeeze(1)
-            
+
             input_ids = torch.cat([input_ids, next_token.unsqueeze(-1)], dim=-1)
 
         return input_ids
